tests/wpt/tests/css/css-text/parsing/text-align-last-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Text Module Test: parsing text-align-last with valid values</title>
<link rel="help" href="https://www.w3.org/TR/css-text-3/#propdef-text-align-last">
<meta name="assert" content="text-align-last supports the full grammar 'auto | start | end | left | right | center | justify | match-parent'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("text-align-last", "auto");
test_valid_value("text-align-last", "start");
test_valid_value("text-align-last", "end");
test_valid_value("text-align-last", "left");
test_valid_value("text-align-last", "right");
test_valid_value("text-align-last", "center");
test_valid_value("text-align-last", "justify");
test_valid_value("text-align-last", "match-parent");
</script>
</body>
</html>
